[FIXME] user/linux-volla-vidofnir: new package
Reboots after 75 seconds, so seems to get to initramfs at least but due
to no USB and a panic call it reboots.

Tbf the way I'm testing any of these kernels has been replace kernel
Image* inside boot.img and that's it. Need to experiment with adding
proper kernel modules from the build in there as well..